Add infer_input_bounds(vector<int>)
Add a variant of infer_input_bounds() that takes an explicit vector of int, rather than the up-to-4-int version that is a holdover from the buffer_t days; deprecate the old version; convert all existing code to use the new one.

There are two warts here, unfortunately: in theory we should be able to just wrap the ints in {braces}, but in practice that doesn't work for zero- or one-arg lists:
- zero-arg vectors are ambiguous wrt other overloads of the method, and the compiler will select a deprecated one
- one-arg vectors are broken in C++ with brace initialization, as a single-integer arg will invoke the "create a vector of this length" ctor rather than "create a vector of length one with this value" ctor.

These two warts leave me with a bad taste in my mouth; the 'new' approach of always a list of sizes, rather than a sorta-vararg approach, is clearly better, and yet, the ugly code in some places is ugly.